Conflicts:
	.gitignore
	cpp/cmd/5tt2gmwmi.cpp
	cpp/cmd/5tt2vis.cpp
	cpp/cmd/5ttcheck.cpp
	cpp/cmd/5ttedit.cpp
	cpp/cmd/afdconnectivity.cpp
	cpp/cmd/amp2response.cpp
	cpp/cmd/amp2sh.cpp
	cpp/cmd/connectome2tck.cpp
	cpp/cmd/connectomeedit.cpp
	cpp/cmd/connectomestats.cpp
	cpp/cmd/dirflip.cpp
	cpp/cmd/dirgen.cpp
	cpp/cmd/dirmerge.cpp
	cpp/cmd/dirorder.cpp
	cpp/cmd/dirrotate.cpp
	cpp/cmd/dirsplit.cpp
	cpp/cmd/dirstat.cpp
	cpp/cmd/dwi2adc.cpp
	cpp/cmd/dwi2fod.cpp
	cpp/cmd/dwi2tensor.cpp
	cpp/cmd/dwidenoise.cpp
	cpp/cmd/dwiextract.cpp
	cpp/cmd/fixel2peaks.cpp
	cpp/cmd/fixel2sh.cpp
	cpp/cmd/fixel2tsf.cpp
	cpp/cmd/fixel2voxel.cpp
	cpp/cmd/fixelcfestats.cpp
	cpp/cmd/fixelconnectivity.cpp
	cpp/cmd/fixelconvert.cpp
	cpp/cmd/fixelcorrespondence.cpp
	cpp/cmd/fixelcrop.cpp
	cpp/cmd/fixelfilter.cpp
	cpp/cmd/fixelreorient.cpp
	cpp/cmd/fod2fixel.cpp
	cpp/cmd/label2colour.cpp
	cpp/cmd/label2mesh.cpp
	cpp/cmd/labelconvert.cpp
	cpp/cmd/labelstats.cpp
	cpp/cmd/maskfilter.cpp
	cpp/cmd/mesh2voxel.cpp
	cpp/cmd/mrclusterstats.cpp
	cpp/cmd/mrconvert.cpp
	cpp/cmd/mrdegibbs.cpp
	cpp/cmd/mredit.cpp
	cpp/cmd/mrfilter.cpp
	cpp/cmd/mrgrid.cpp
	cpp/cmd/mrinfo.cpp
	cpp/cmd/mrmath.cpp
	cpp/cmd/mrmetric.cpp
	cpp/cmd/mrregister.cpp
	cpp/cmd/mrthreshold.cpp
	cpp/cmd/mrtransform.cpp
	cpp/cmd/mtnormalise.cpp
	cpp/cmd/peaks2amp.cpp
	cpp/cmd/peaks2fixel.cpp
	cpp/cmd/sh2amp.cpp
	cpp/cmd/sh2peaks.cpp
	cpp/cmd/sh2power.cpp
	cpp/cmd/shbasis.cpp
	cpp/cmd/shconv.cpp
	cpp/cmd/tck2connectome.cpp
	cpp/cmd/tck2fixel.cpp
	cpp/cmd/tckconvert.cpp
	cpp/cmd/tckdfc.cpp
	cpp/cmd/tckgen.cpp
	cpp/cmd/tckglobal.cpp
	cpp/cmd/tckmap.cpp
	cpp/cmd/tckresample.cpp
	cpp/cmd/tcksample.cpp
	cpp/cmd/tcktransform.cpp
	cpp/cmd/transformcalc.cpp
	cpp/cmd/transformcompose.cpp
	cpp/cmd/transformconvert.cpp
	cpp/cmd/tsfdivide.cpp
	cpp/cmd/tsfmult.cpp
	cpp/cmd/tsfsmooth.cpp
	cpp/cmd/tsfvalidate.cpp
	cpp/cmd/vectorstats.cpp
	cpp/cmd/voxel2mesh.cpp
	cpp/cmd/warp2metric.cpp
	cpp/cmd/warpconvert.cpp
	cpp/cmd/warpcorrect.cpp
	cpp/cmd/warpinvert.cpp
	cpp/core/algo/histogram.cpp
	cpp/core/algo/histogram.h
	cpp/core/app.cpp
	cpp/core/app.h
	cpp/core/connectome/connectome.cpp
	cpp/core/connectome/lut.cpp
	cpp/core/connectome/lut.h
	cpp/core/dwi/directions/file.cpp
	cpp/core/dwi/directions/file.h
	cpp/core/dwi/directions/set.h
	cpp/core/dwi/gradient.cpp
	cpp/core/dwi/gradient.h
	cpp/core/dwi/sdeconv/csd.h
	cpp/core/dwi/sdeconv/msmt_csd.h
	cpp/core/dwi/tractography/ACT/act.cpp
	cpp/core/dwi/tractography/ACT/shared.h
	cpp/core/dwi/tractography/GT/gt.h
	cpp/core/dwi/tractography/SIFT/model.h
	cpp/core/dwi/tractography/SIFT/model_base.h
	cpp/core/dwi/tractography/SIFT/output.h
	cpp/core/dwi/tractography/SIFT/sifter.cpp
	cpp/core/dwi/tractography/SIFT/sifter.h
	cpp/core/dwi/tractography/SIFT2/tckfactor.cpp
	cpp/core/dwi/tractography/SIFT2/tckfactor.h
	cpp/core/dwi/tractography/algorithms/fact.h
	cpp/core/dwi/tractography/algorithms/iFOD1.h
	cpp/core/dwi/tractography/algorithms/iFOD2.h
	cpp/core/dwi/tractography/algorithms/nulldist.h
	cpp/core/dwi/tractography/algorithms/sd_stream.h
	cpp/core/dwi/tractography/algorithms/seedtest.h
	cpp/core/dwi/tractography/algorithms/tensor_det.h
	cpp/core/dwi/tractography/algorithms/tensor_prob.h
	cpp/core/dwi/tractography/connectome/extract.cpp
	cpp/core/dwi/tractography/connectome/extract.h
	cpp/core/dwi/tractography/connectome/matrix.cpp
	cpp/core/dwi/tractography/connectome/matrix.h
	cpp/core/dwi/tractography/connectome/metric.h
	cpp/core/dwi/tractography/editing/receiver.h
	cpp/core/dwi/tractography/file.h
	cpp/core/dwi/tractography/file_base.cpp
	cpp/core/dwi/tractography/file_base.h
	cpp/core/dwi/tractography/mapping/buffer_scratch_dump.h
	cpp/core/dwi/tractography/mapping/mapper.cpp
	cpp/core/dwi/tractography/mapping/mapper.h
	cpp/core/dwi/tractography/mapping/mapping.cpp
	cpp/core/dwi/tractography/mapping/mapping.h
	cpp/core/dwi/tractography/mapping/writer.h
	cpp/core/dwi/tractography/roi.cpp
	cpp/core/dwi/tractography/roi.h
	cpp/core/dwi/tractography/scalar_file.h
	cpp/core/dwi/tractography/seeding/base.h
	cpp/core/dwi/tractography/seeding/basic.cpp
	cpp/core/dwi/tractography/seeding/basic.h
	cpp/core/dwi/tractography/seeding/dynamic.h
	cpp/core/dwi/tractography/seeding/gmwmi.cpp
	cpp/core/dwi/tractography/seeding/gmwmi.h
	cpp/core/dwi/tractography/seeding/seeding.cpp
	cpp/core/dwi/tractography/tracking/exec.h
	cpp/core/dwi/tractography/tracking/shared.cpp
	cpp/core/dwi/tractography/tracking/shared.h
	cpp/core/dwi/tractography/tracking/write_kernel.h
	cpp/core/file/config.cpp
	cpp/core/file/copy.h
	cpp/core/file/dicom/element.cpp
	cpp/core/file/dicom/element.h
	cpp/core/file/dicom/image.cpp
	cpp/core/file/dicom/image.h
	cpp/core/file/dicom/quick_scan.cpp
	cpp/core/file/dicom/quick_scan.h
	cpp/core/file/dicom/tree.cpp
	cpp/core/file/dicom/tree.h
	cpp/core/file/entry.h
	cpp/core/file/gz.h
	cpp/core/file/json_utils.cpp
	cpp/core/file/json_utils.h
	cpp/core/file/key_value.cpp
	cpp/core/file/key_value.h
	cpp/core/file/matrix.h
	cpp/core/file/mgh.h
	cpp/core/file/mmap.cpp
	cpp/core/file/name_parser.cpp
	cpp/core/file/name_parser.h
	cpp/core/file/nifti_utils.cpp
	cpp/core/file/nifti_utils.h
	cpp/core/file/npy.cpp
	cpp/core/file/npy.h
	cpp/core/file/ofstream.cpp
	cpp/core/file/ofstream.h
	cpp/core/file/path.h
	cpp/core/file/png.cpp
	cpp/core/file/png.h
	cpp/core/file/utils.cpp
	cpp/core/file/utils.h
	cpp/core/fixel/filter/connect.cpp
	cpp/core/fixel/filter/smooth.cpp
	cpp/core/fixel/helpers.h
	cpp/core/fixel/matrix.cpp
	cpp/core/fixel/matrix.h
	cpp/core/formats/mgh.cpp
	cpp/core/formats/mgz.cpp
	cpp/core/formats/mri.cpp
	cpp/core/formats/mrtrix_sparse_legacy.cpp
	cpp/core/formats/mrtrix_utils.cpp
	cpp/core/formats/mrtrix_utils.h
	cpp/core/formats/par.cpp
	cpp/core/formats/xds.cpp
	cpp/core/header.cpp
	cpp/core/header.h
	cpp/core/image.h
	cpp/core/image_io/default.cpp
	cpp/core/image_io/gz.cpp
	cpp/core/image_io/pipe.cpp
	cpp/core/image_io/sparse.cpp
	cpp/core/math/stats/glm.cpp
	cpp/core/math/stats/glm.h
	cpp/core/math/stats/import.h
	cpp/core/math/stats/shuffle.cpp
	cpp/core/math/stats/shuffle.h
	cpp/core/phase_encoding.cpp
	cpp/core/phase_encoding.h
	cpp/core/registration/multi_contrast.h
	cpp/core/signal_handler.cpp
	cpp/core/signal_handler.h
	cpp/core/surface/freesurfer.cpp
	cpp/core/surface/freesurfer.h
	cpp/core/surface/mesh.cpp
	cpp/core/surface/mesh.h
	cpp/core/surface/mesh_multi.cpp
	cpp/core/surface/mesh_multi.h
	cpp/core/surface/scalar.cpp
	cpp/core/surface/scalar.h
	cpp/gui/dialog/file.cpp
	cpp/gui/dialog/file.h
	cpp/gui/dialog/image_properties.cpp
	cpp/gui/dwi/render_frame.cpp
	cpp/gui/dwi/render_frame.h
	cpp/gui/mrview/displayable.cpp
	cpp/gui/mrview/displayable.h
	cpp/gui/mrview/tool/connectome/connectome.cpp
	cpp/gui/mrview/tool/connectome/file_data_vector.cpp
	cpp/gui/mrview/tool/connectome/file_data_vector.h
	cpp/gui/mrview/tool/fixel/base_fixel.cpp
	cpp/gui/mrview/tool/fixel/base_fixel.h
	cpp/gui/mrview/tool/fixel/directory.cpp
	cpp/gui/mrview/tool/fixel/directory.h
	cpp/gui/mrview/tool/fixel/fixel.cpp
	cpp/gui/mrview/tool/fixel/image4D.h
	cpp/gui/mrview/tool/fixel/legacy.h
	cpp/gui/mrview/tool/odf/item.cpp
	cpp/gui/mrview/tool/overlay.cpp
	cpp/gui/mrview/tool/tractography/tractogram.cpp
	cpp/gui/mrview/tool/tractography/tractogram.h
	cpp/gui/shview/render_window.cpp
	cpp/gui/shview/render_window.h
	testing/tools/testing_diff_mesh.cpp
	testing/tools/testing_npywrite.cpp

+++ cpp/core/dwi/tractography/connectome/extract.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2008-2025 the MRtrix3 contributors.
+/* Copyright (c) 2008-2026 the MRtrix3 contributors.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -55,7 +55,7 @@ public:
                   const std::vector<node_t> &,
                   const bool,
                   const node_t,
-                  const std::vector<Eigen::Vector3f> &);
+                  const std::vector<Eigen::Vector3d> &);
 
   bool operator()(const Tractography::Connectome::Streamline_nodepair &);
   bool operator()(const Tractography::Connectome::Streamline_nodelist &);
